feat: add SVG padding prop to ShapeEditor
Padding _inside_ the `<svg>`, in px. Useful for displaying SVG content that
would normally get clipped off by the edge of the SVG element. Can also help
when providing contextual information (e.g., the surrounding parts of an image)
while keeping the 0,0 coordinates on the part of the image you are focused on,
avoiding having to convert the coordinates back-and-forth. Can be set via a
number to apply to all edges, or as an object in the form:
`{ top: number, right: number, bottom: number, left: number }` to apply
per-side.