fix: do not allow edit actions to be initiated by non-left clicks
Previously, right clicking on a DrawLayer, SelectionLayer or wrapped
shape would result in an edit action being initiated, immediately
followed by the context menu appearing. Clicking out of the context menu
would then leave the editor in an in-between state, where the mouse
button was not pressed, but the shape edit continued on as if it was.